Compiles on Vogon.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@786 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Items/AsteroidShip.cpp
@@ -135,8 +135,8 @@ AsteroidShip::AsteroidShip(const GameState* _gameState) :
    shooter = new ShootingAI(this);
    flyingAI = new FlyingAI(this);
 
-   color1 = randdouble();
-   color2 = randdouble();
+   color1 = (float) randdouble();
+   color2 = (float) randdouble();
 }
 
 /**
